create new backdoor
1. Refactor Backdoor folders
2. Create new HTTP backdoor
3. create HowToUse md files
4. delete standalones
