+++ malwares/reverse_backdoor/TCP/HowToUse.md
@@ -1,0 +1,32 @@
+
+## Change Values according to need
+- change ip and port to your desired values in listener.py and reverse_backdoor.py
+
+- run listener.py on attackers machine.
+
+- run backdoor.py on victims machine.
+
+## Create a executable\standalone
+- pip3 install pyinstaller
+
+- Creating executable with console: 
+    ```bash
+    $ pyinstaller python_file.py --onefile
+    ```
+
+- Creating executable without console: 
+    ```bash
+    $ pyinstaller python_file.py --onefile --noconsole
+    ```
+
+
+> note : noconsole works when we're not using stream like stdin, stdout, stderr, etc. 
+> If we're using STDI/O streams then we have to set then to DEVNULL = open(os.devnull, 'wb') then set IOstreams to DEVNULL
+> if using subprocess.check_output then use subprocess.check_output(command, stderr=subprocess.DEVNULL, stdin=subprocess.DEVNULL).
+> here stdio is being handled by check_output. so no need to handle stdio.
+
+## Create windows executable on linux
+- Install Wine 
+- Download Python for windows 
+- Install Downloaded python for windows using wine on linux 
+- the other commands remain the same to create executable.
