--- malwares/reverse_backdoor/howtouse.txt
@@ -1,20 +1,0 @@
-1. change ip and port to your desired values in listener.py and reverse_backdoor.py
-2. run listener.py on attackers machine.
-3. run backdoor.py on victims machine.
-
-# Creating a executable\standalone
-1. pip3 install pyinstaller
-2. Creating executable with console: 
-    $ pyinstaller python_file.py --onefile
-3. Creating executable without console: 
-    $ pyinstaller python_file.py --onefile --noconsole
-# note : noconsole works when we're not using stream like stdin, stdout, stderr, etc. 
-# If we're using STDI/O streams then we have to set then to DEVNULL = open(os.devnull, 'wb') then set IOstreams to DEVNULL
-# if using subprocess.check_output then use subprocess.check_output(command, stderr=subprocess.DEVNULL, stdin=subprocess.DEVNULL).
-# here stdio is being handled by check_output. so no need to handle stdio.
-
-# Creating windows executable on linux
-1. Install Wine 
-2. Download Python for windows 
-3. Install Downloaded python for windows using wine on linux 
-4. the other commands remain the same to create executable.
